+++ CHANGELOG.rst
@@ -12,6 +12,7 @@ Versioning <http://semver.org/>`__. The changelog format is inspired by
 - Update the [vasp] extra to handle some compatibility breaks between
   dependencies and different Python versions.
 - Fix some incorrect values in Al k-alpha XPS cross-sections
+- BUGFIX: Pymatgen CompleteDOS was not correctly accepted by galore.process_pdos()
 - Implement previously ineffective "offset" option in
   galore.plot.plot_pdos(), add a matching option to
   galore.plot.plot_tdos()